requirements.txt
numpy==1.26.4
pyarrow==14.0.2
pandas==2.2.3
datasets==2.15.0
numba==0.59.1
scipy<1.15
huggingface-hub==0.23.2
transformers==4.36.2
tokenizers==0.15.2
torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
Cython
pybind11
onnxruntime
torchcodec
sounddevice